Edit and delete posting

+++ app/src/main/java/com/example/tocarcar/entity/Posting.kt
@@ -9,7 +9,7 @@ data class Posting(
     var email: String = "",
     var dateFrom: String = "",
     var dateTo: String = "",
-    var rentPerDay: Float = 0.0F,
+    var rentPerDay: Double = 0.0,
     var isApproved: Int = 0,
     var isBooked: Int = 0,
     var bookedBy: String = ""
